UPdated SDK, Stroop  Loading, Movement
Added movement for shoulder  tracking.
Fixed the stroop  loading. It was pointing to  reources in the  Assets forlder and I was trying to use Resources.Load on a folder in Assets, not in Resources. Which I guess is wrong. Don't really understand what Resources distinction means, but maybe those assets are builts with the project?
Stroop Works!
Math should too
Updated to v65  of the sdk