Updated and not great
the hand interaction offsets are weird now, you need to be at the knuckle of the hand before it works, grab now requires a full hand, all the hand tracking is worse and offset. the world slides with you slightly as you walk which isn't gonna do it. reverting while keeping the script and material changes